secret_manager/requirements.txt
google-auth
google-cloud-secret-manager